remove env.cmd from gitignore.
I think this was mistakenly added to gitignore in an incorrect attempt to prevent user specific updates from being committed. This will not work because gitignore only works if a file is not already tracked. To get the desired behaviour, we would need to create a separate template file which creates an env.cmd when called.

In the meantime. Just put these tools in the location specified in this file. Or otherwise, do not commit changes.